+++ packages/basepaint/src/Basepaint.ts
@@ -9,7 +9,7 @@ import { BASE_CHAIN_ID, CONTRACT_ADDRESS, MINT_ABI } from './constants'
 export const mint = async (
   mint: MintActionParams,
 ): Promise<TransactionFilter> => {
-  const { address, tokenId, quantity } = mint
+  const { address, tokenId, amount } = mint
 
   return compressJson({
     chainId: BASE_CHAIN_ID,
@@ -17,7 +17,7 @@ export const mint = async (
     input: {
       $abi: MINT_ABI,
       day: tokenId,
-      count: quantity,
+      count: amount,
     },
   })
 }