+++ packages/camelot/src/Camelot.ts
@@ -2,52 +2,112 @@ import {
   type TransactionFilter,
   type SwapActionParams,
   compressJson,
-  type FilterOperator,
 } from '@rabbitholegg/questdk'
 import { type Address } from 'viem'
 import { CHAIN_ID_ARRAY, ARBITRUM_CHAIN_ID } from './chain-ids'
+import { DEFAULT_TOKEN_LIST } from './contract-addresses'
+import { buildPathQuery, Tokens } from './utils'
+import { CAMELOT_ABI, PARASWAP_ABI } from './abi'
 import {
-  DEFAULT_TOKEN_LIST_URL,
-  WETH_ADDRESS_ARBITRUM,
+  CAMELOT_ROUTER,
+  PARASWAP_ROUTER,
+  ETH_ADDRESS,
 } from './contract-addresses'
-import { CAMELOT_ABI } from './abi'
-import { CAMELOT_ROUTER } from './contract-addresses'
 
-export const buildPathQuery = (tokenIn?: string, tokenOut?: string) => {
-  // v2 paths are formatted as [<token>, <token>]
-  const conditions: FilterOperator[] = []
-
-  if (tokenIn) {
-    conditions.push({ $first: tokenIn })
-  }
-
-  if (tokenOut) {
-    conditions.push({ $last: tokenOut })
-  }
-
-  return {
-    $and: conditions,
-  }
+const isValidContractAddress = (address: Address) => {
+  return (
+    address?.toLowerCase() === CAMELOT_ROUTER.toLowerCase() ||
+    address?.toLowerCase() === PARASWAP_ROUTER.toLowerCase()
+  )
 }
 
 export const swap = async (
   swap: SwapActionParams,
 ): Promise<TransactionFilter> => {
-  const { chainId, contractAddress, tokenIn, tokenOut, amountOut, recipient } =
-    swap
-  const amountIn = tokenIn === WETH_ADDRESS_ARBITRUM ? swap.amountIn : undefined
+  const {
+    chainId,
+    contractAddress,
+    tokenIn,
+    tokenOut,
+    amountIn,
+    amountOut,
+    recipient,
+  } = swap
 
-  // We always want to return a compressed JSON object which we'll transform into a TransactionFilter
+  const ethUsedIn = tokenIn === Tokens.ETH
+  const ethUsedOut = tokenOut === Tokens.ETH
+
+  if (contractAddress && !isValidContractAddress(contractAddress)) {
+    throw new Error('Invalid Contract Address')
+  }
+
   return compressJson({
-    chainId: chainId, // The chainId of the source chain
-    to: contractAddress || CAMELOT_ROUTER, // The contract address of the bridge
+    chainId: chainId,
+    to: { $or: [CAMELOT_ROUTER.toLowerCase(), PARASWAP_ROUTER.toLowerCase()] },
+    value: ethUsedIn ? amountIn : undefined,
     input: {
-      $abi: CAMELOT_ABI, // The ABI of the bridge
-      to: recipient, // The recipient of the swap
-      path: buildPathQuery(tokenIn, tokenOut), // The path of the swap
-      amountOutMin: amountOut, // The minimum amount of tokens to receive
-      amountIn: amountIn, // The amount of tokens to send
-    }, // The input object is where we'll put the ABI and the parameters
+      $abi: [...CAMELOT_ABI, ...PARASWAP_ABI],
+      $or: [
+        {
+          // camelotV2 swap
+          to: recipient,
+          path: buildPathQuery(ethUsedIn ? Tokens.WETH : tokenIn, tokenOut),
+          amountOutMin: amountOut,
+          amountIn: ethUsedIn ? undefined : amountIn,
+        },
+        {
+          // simpleswap, directUniV3Swap, directCurveSwap
+          data: {
+            fromToken: ethUsedIn ? ETH_ADDRESS : tokenIn,
+            fromAmount: amountIn,
+            toAmount: amountOut,
+            toToken: ethUsedOut ? ETH_ADDRESS : tokenOut,
+            partner: '0x353D2d14Bb674892910685520Ac040f560CcBC06',
+          },
+        },
+        {
+          // multiswap
+          data: {
+            fromToken: ethUsedIn ? ETH_ADDRESS : tokenIn,
+            fromAmount: amountIn,
+            toAmount: amountOut,
+            path: {
+              $last: {
+                to: ethUsedOut ? ETH_ADDRESS : tokenOut,
+              },
+            },
+            partner: '0x353D2d14Bb674892910685520Ac040f560CcBC06',
+          },
+        },
+        {
+          // megaswap
+          data: {
+            fromToken: ethUsedIn ? ETH_ADDRESS : tokenIn,
+            fromAmount: amountIn,
+            toAmount: amountOut,
+            path: {
+              $last: {
+                path: {
+                  $last: {
+                    to: ethUsedOut ? ETH_ADDRESS : tokenOut,
+                  },
+                },
+              },
+            },
+            partner: '0x353D2d14Bb674892910685520Ac040f560CcBC06',
+          },
+        },
+        {
+          // directBalancerV2
+          data: {
+            assets: buildPathQuery(tokenIn, tokenOut),
+            fromAmount: amountIn,
+            toAmount: amountOut,
+            partner: '0x353D2d14Bb674892910685520Ac040f560CcBC06',
+          },
+        },
+      ],
+    },
   })
 }
 
@@ -55,7 +115,7 @@ export const getSupportedTokenAddresses = async (
   _chainId: number,
 ): Promise<Address[]> => {
   // Only return supported tokens for ARBITRUM_CHAIN_ID
-  return _chainId === ARBITRUM_CHAIN_ID ? DEFAULT_TOKEN_LIST_URL : []
+  return _chainId === ARBITRUM_CHAIN_ID ? DEFAULT_TOKEN_LIST : []
 }
 
 export const getSupportedChainIds = async (): Promise<number[]> => {