+++ packages/okutrade/src/index.ts
@@ -1,0 +1,19 @@
+import {
+  type IActionPlugin,
+  PluginActionNotImplementedError,
+} from '@rabbitholegg/questdk'
+
+import {
+  swap,
+  getSupportedChainIds,
+  getSupportedTokenAddresses,
+} from './OkuTrade.js'
+
+export const OkuTrade: IActionPlugin = {
+  pluginId: 'okutrade',
+  getSupportedTokenAddresses,
+  getSupportedChainIds,
+  swap,
+  bridge: async () => new PluginActionNotImplementedError(),
+  mint: async () => new PluginActionNotImplementedError(),
+}