+++ packages/registry/src/index.ts
@@ -6,6 +6,7 @@ import {
   PluginActionNotImplementedError,
   type MintActionParams,
   type SwapActionParams,
+  type QuestActionParams,
   type DelegateActionParams,
   type TransactionFilter,
   type StakeActionParams,
@@ -25,7 +26,9 @@ import { Tally } from '@rabbitholegg/questdk-plugin-tally'
 import { BasePaint } from '@rabbitholegg/questdk-plugin-basepaint'
 import { Hyphen } from '@rabbitholegg/questdk-plugin-hyphen'
 import { Paraswap } from '@rabbitholegg/questdk-plugin-paraswap'
+import { Rabbithole } from '@rabbitholegg/questdk-plugin-rabbithole'
 import { Symbiosis } from '@rabbitholegg/questdk-plugin-symbiosis'
+import { OkuTrade } from '@rabbitholegg/questdk-plugin-okutrade'
 import { Synapse } from '@rabbitholegg/questdk-plugin-synapse'
 import { ENTRYPOINT } from './contract-addresses'
 
@@ -44,8 +47,10 @@ export const plugins: Record<string, IActionPlugin> = {
   [BasePaint.pluginId]: BasePaint,
   [Hyphen.pluginId]: Hyphen,
   [Paraswap.pluginId]: Paraswap,
+  [Rabbithole.pluginId]: Rabbithole,
   [Symbiosis.pluginId]: Symbiosis,
   [Synapse.pluginId]: Synapse,
+  [OkuTrade.pluginId]: OkuTrade,
 }
 
 export const getPlugin = (pluginId: string) => {
@@ -72,6 +77,11 @@ export const executePlugin = (
       if (plugin.delegate === undefined) {
         return Promise.reject(new PluginActionNotImplementedError())
       } else return plugin.delegate(params as unknown as DelegateActionParams)
+    }
+    case ActionType.Quest: {
+      if (plugin.quest === undefined) {
+        return Promise.reject(new PluginActionNotImplementedError())
+      } else return plugin.quest(params as unknown as QuestActionParams)
     }
     case ActionType.Stake: {
       if (plugin.stake === undefined) {